+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support calling `/api/v1/accounts/search` with `resolve` parameter.
+- Created `/api/v1/accounts/aliases/all` API endpoint.
 
 ### Changed